requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
